social_network/requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
pip==22.0.2
setuptools==59.6.0
sqlparse==0.5.1
typing_extensions==4.12.2
